Change task_1

+++ Practice_4/Practice_4.Task_1/Program.cs
@@ -5,7 +5,7 @@ namespace Practice_4.Task_1
     internal class Program
     {
 
-        static double y1(double x)
+        static int y1(int x)
         {
             return ((5 * x - 7) / (4 * x + 8));
         }
@@ -20,10 +20,10 @@ namespace Practice_4.Task_1
             try
             {
                 Console.WriteLine("Введите значение Х1: ");
-                double x1 = double.Parse(Console.ReadLine());
+                int x1 = int.Parse(Console.ReadLine());
                 Console.WriteLine($"Результат формулы y1: " + y1(x1));
                 Console.WriteLine("Введите значение Х2: ");
-                double x2 = double.Parse(Console.ReadLine());
+                int x2 = int.Parse(Console.ReadLine());
                 Console.WriteLine($"Результат формулы y2: " + y2(x2));
             }
             catch (FormatException ex)
